+++ src/components/organisms/Layout/StudentsTable/StyledTableCell.tsx
@@ -1,0 +1,13 @@
+import { TableCell, styled, tableCellClasses } from "@mui/material";
+
+const StyledTableCell = styled(TableCell)(({ theme }) => ({
+  [`&.${tableCellClasses.head}`]: {
+    backgroundColor: theme.palette.info.light,
+    color: theme.palette.common.white,
+  },
+  [`&.${tableCellClasses.body}`]: {
+    fontSize: 14,
+  },
+}));
+
+export default StyledTableCell
