+++ src/components/organisms/Layout/StudentsTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './StudentsTable'
